+++ smok/extras/sensor_write_database/in_memory.py
@@ -6,19 +6,19 @@ from ...sensor import SensorWriteEvent
 
 class InMemorySensorWriteDatabase(BaseSensorWriteDatabase, Monitor):
     @Monitor.synchronized
-    def on_synced(self, event: SensorWriteEvent):
+    def on_synced_sw(self, event: SensorWriteEvent):
         self.events.remove(event)
 
     @Monitor.synchronized
-    def on_sync_failed(self, event: SensorWriteEvent):
+    def on_sync_sw_failed(self, event: SensorWriteEvent):
         pass
 
     @Monitor.synchronized
-    def add(self, event: SensorWriteEvent):
+    def add_sw(self, event: SensorWriteEvent):
         self.events.add(event)
 
     @Monitor.synchronized
-    def get_sync(self) -> BaseSensorWriteSynchronization:
+    def get_sw_sync(self) -> BaseSensorWriteSynchronization:
         return BaseSensorWriteSynchronization(list(self.events), self)
 
     def __init__(self):